# AssemblyScanner improvements