--- a/src/NServiceBus.Core.Tests/AssemblyScanner/When_exclusion_predicate_is_used.cs
+++ b/src/NServiceBus.Core.Tests/AssemblyScanner/When_exclusion_predicate_is_used.cs
@@ -16,9 +16,9 @@
         {
             var results = new AssemblyScanner(Path.Combine(TestContext.CurrentContext.TestDirectory, "TestDlls"))
             {
-                AssembliesToSkip = new HashSet<string>(StringComparer.OrdinalIgnoreCase)
+                AssembliesToSkip = new List<string>
                     {
-                        "dotNet"
+                        "dotNet.dll"
                     },
                 ScanAppDomainAssemblies = false
             }