--- a/src/NServiceBus.Core.Tests/AssemblyScanner/When_validating_assemblies.cs
+++ b/src/NServiceBus.Core.Tests/AssemblyScanner/When_validating_assemblies.cs
@@ -10,12 +10,11 @@
         [Test]
         public void Should_not_validate_system_assemblies()
         {
-            var validator = new AssemblyValidator();
             var systemAssemblies = AppDomain.CurrentDomain.GetAssemblies().Where(a => a.FullName.Contains("System"));
 
             foreach (var assembly in systemAssemblies)
             {
-                validator.ValidateAssemblyFile(assembly.Location, out var shouldLoad, out var reason);
+                AssemblyValidator.ValidateAssemblyFile(assembly.Location, out var shouldLoad, out var reason);
 
                 Assert.IsFalse(shouldLoad, $"Should not validate {assembly.FullName}");
                 Assert.That(reason == "File is a .NET runtime assembly.");
@@ -25,9 +24,7 @@
         [Test]
         public void Should_validate_NServiceBus_Core_assembly()
         {
-            var validator = new AssemblyValidator();
-
-            validator.ValidateAssemblyFile(typeof(EndpointConfiguration).Assembly.Location, out var shouldLoad, out var reason);
+            AssemblyValidator.ValidateAssemblyFile(typeof(EndpointConfiguration).Assembly.Location, out var shouldLoad, out var reason);
 
             Assert.IsTrue(shouldLoad);
             Assert.That(reason == "File is a .NET assembly.");
@@ -36,9 +33,7 @@
         [Test]
         public void Should_validate_non_system_assemblies()
         {
-            var validator = new AssemblyValidator();
-
-            validator.ValidateAssemblyFile(GetType().Assembly.Location, out var shouldLoad, out var reason);
+            AssemblyValidator.ValidateAssemblyFile(GetType().Assembly.Location, out var shouldLoad, out var reason);
 
             Assert.IsTrue(shouldLoad);
             Assert.That(reason == "File is a .NET assembly.");