--- a/src/NServiceBus.Core.Tests/Config/When_scanning_assemblies.cs
+++ b/src/NServiceBus.Core.Tests/Config/When_scanning_assemblies.cs
@@ -1,6 +1,5 @@
 namespace NServiceBus.Core.Tests.Config
 {
-    using System;
     using System.Collections.Generic;
     using System.Linq;
     using System.Reflection;