… src/NServiceBus.Core/Hosting/Helpers/AssemblyValidator.cs