Fixes https://github.com/Particular/NServiceBus/issues/6820

Improves the assembly scanner performance in "regular cases" slightly and in cases when the exclusion list is larger, up to 4 times.

I was doing some investigations for a customer that mentioned they are seeing slow startup times in their rather large solution. Once I was able to get my hands on the profiler data it was apparent that the `IsExcluded` check is the culprit. 

## Scenarios

I have created a [solution generator](https://github.com/danielmarbach/NServiceBusAssemblyScanning) that allows simulating arbitrary solution sizes. As a reference point, I created a solution with roughly 960 projects and more than 140K types. 

### Without Exclusions

#### Before

![image](https://github.com/Particular/NServiceBus/assets/174258/79515641-5f74-4e17-aec7-a0061eb7f4ac)


#### After

![image](https://github.com/Particular/NServiceBus/assets/174258/8c75e2f2-d50f-4447-9b3f-7ad8c17d4b64)


### With a larger number of exclusions

#### Before

![image](https://github.com/Particular/NServiceBus/assets/174258/d832902b-0279-4d58-8d40-22a1deabca20)

Here is the excerpt from the customer callstack

![image](https://github.com/Particular/NServiceBus/assets/174258/00c142f5-afba-4d49-b19f-276e2799c635)


#### After

![image](https://github.com/Particular/NServiceBus/assets/174258/c3ff4de9-3b23-4e58-9e53-7ac51e463365)

## Benchmarks

### Attribute.GetCustomAttribute

|         Method |       Mean |     Error |    StdDev |     Median | Ratio | RatioSD |
|--------------- |-----------:|----------:|----------:|-----------:|------:|--------:|
|           Type | 1,148.2 ns | 135.51 ns | 386.63 ns | 1,010.9 ns |  1.00 |    0.00 |
| AttributeBased |   804.7 ns |  12.05 ns |  11.27 ns |   801.7 ns |  0.99 |    0.20 |

### Matching Strings

|                Method |            Input |      Mean |    Error |   StdDev | Ratio |  Gen 0 | Allocated |
|---------------------- |----------------- |----------:|---------:|---------:|------:|-------:|----------:|
|         **MatchOriginal** |     **SomeAssembly** | **146.94 ns** | **0.815 ns** | **0.680 ns** |  **1.00** | **0.0010** |      **96 B** |
|        MatchOptimized |     SomeAssembly |  28.90 ns | 0.441 ns | 0.368 ns |  0.20 |      - |         - |
| MatchOptimizedHashSet |     SomeAssembly |  17.72 ns | 0.057 ns | 0.047 ns |  0.12 |      - |         - |
|                       |                  |           |          |          |       |        |           |
|         **MatchOriginal** | **SomeAssembly.Dll** | **149.63 ns** | **2.527 ns** | **2.363 ns** |  **1.00** | **0.0024** |     **208 B** |
|        MatchOptimized | SomeAssembly.Dll |  59.56 ns | 0.504 ns | 0.447 ns |  0.40 | 0.0011 |      96 B |
| MatchOptimizedHashSet | SomeAssembly.Dll |  11.01 ns | 0.106 ns | 0.089 ns |  0.07 |      - |         - |
|                       |                  |           |          |          |       |        |           |
|         **MatchOriginal** | **SomeAssembly.Exe** | **184.95 ns** | **1.662 ns** | **1.473 ns** |  **1.00** | **0.0024** |     **208 B** |
|        MatchOptimized | SomeAssembly.Exe |  72.69 ns | 0.963 ns | 0.901 ns |  0.39 | 0.0011 |      96 B |
| MatchOptimizedHashSet | SomeAssembly.Exe |  10.87 ns | 0.038 ns | 0.036 ns |  0.06 |      - |         - |


### PublicKeyValidation

| Method |       Mean |    Error |   StdDev | Ratio |  Gen 0 | Allocated |
|------- |-----------:|---------:|---------:|------:|-------:|----------:|
| Before | 1,157.2 ns | 13.39 ns | 12.53 ns |  1.00 | 0.0687 |     432 B |
|  After |   842.8 ns |  7.07 ns |  5.90 ns |  0.73 | 0.0277 |     176 B |