@danielmarbach 2023-08-25
Once the PR is reviewed, I would want to patch https://github.com/Particular/NServiceBus/pull/6814/commits/92da68a0e01fefd021b9d481d0f7377c95d1dba1 on current supported versions